+++ test/java/interbase.java
@@ -79,9 +79,9 @@ class interbase {
 		String	floatvar;
 		String[]	cols;
 		String[]	fields;
-		int	port;
+		short	port;
 		String	socket;
-		int	id;
+		short	id;
 		String	filename;
 		long[]	fieldlens;
 	
@@ -92,7 +92,8 @@ class interbase {
 		}
 	
 		// instantiation
-		SQLRConnection con=new SQLRConnection(args[0],Integer.parseInt(args[1]), 
+		SQLRConnection con=new SQLRConnection(args[0],
+					(short)Integer.parseInt(args[1]), 
 						args[2],args[3],args[4],0,1);
 		SQLRCursor cur=new SQLRCursor(con);
 	
@@ -720,7 +721,7 @@ class interbase {
 	
 		//System.out.println("COMMIT AND ROLLBACK: ");
 		SQLRConnection secondcon=new SQLRConnection(args[0],
-					Integer.parseInt(args[1]), 
+					(short)Integer.parseInt(args[1]), 
 					args[2],args[3],args[4],0,1);
 		SQLRCursor secondcur=new SQLRCursor(secondcon);
 		checkSuccess(secondcur.sendQuery("select count(*) from testtable"),1);
