+++ test/java/postgresql6.java
@@ -74,9 +74,9 @@ class postgresql6 {
 		String	floatvar;
 		String[]	cols;
 		String[]	fields;
-		int	port;
+		short	port;
 		String	socket;
-		int	id;
+		short	id;
 		String	filename;
 		long[]	fieldlens;
 	
@@ -88,7 +88,8 @@ class postgresql6 {
 	
 	
 		// instantiation
-		SQLRConnection con=new SQLRConnection(args[0],Integer.parseInt(args[1]), 
+		SQLRConnection con=new SQLRConnection(args[0],
+					(short)Integer.parseInt(args[1]), 
 						args[2],args[3],args[4],0,1);
 		SQLRCursor cur=new SQLRCursor(con);
 	
@@ -678,7 +679,7 @@ class postgresql6 {
 	
 		System.out.println("COMMIT AND ROLLBACK: ");
 		SQLRConnection secondcon=new SQLRConnection(args[0],
-					Integer.parseInt(args[1]), 
+					(short)Integer.parseInt(args[1]), 
 					args[2],args[3],args[4],0,1);
 		SQLRCursor secondcur=new SQLRCursor(secondcon);
 		checkSuccess(secondcur.sendQuery("select count(*) from testtable"),1);
