+++ test/java/sqlite.java
@@ -74,9 +74,9 @@ class sqlite {
 		String	floatvar;
 		String[]	cols;
 		String[]	fields;
-		int	port;
+		short	port;
 		String	socket;
-		int	id;
+		short	id;
 		String	filename;
 		long[]	fieldlens;
 	
@@ -87,7 +87,8 @@ class sqlite {
 		}
 	
 		// instantiation
-		SQLRConnection con=new SQLRConnection(args[0],Integer.parseInt(args[1]), 
+		SQLRConnection con=new SQLRConnection(args[0],
+					(short)Integer.parseInt(args[1]), 
 						args[2],args[3],args[4],0,1);
 		SQLRCursor cur=new SQLRCursor(con);
 	
@@ -589,7 +590,7 @@ class sqlite {
 
 	    	System.out.println("COMMIT AND ROLLBACK: \n");
 	    	SQLRConnection secondcon=new SQLRConnection(args[0],
-				    	Integer.parseInt(args[1]), 
+				    	(short)Integer.parseInt(args[1]), 
 				    	args[2],args[3],args[4],0,1);
 	    	SQLRCursor secondcur=new SQLRCursor(secondcon);
 	    	checkSuccess(secondcur.sendQuery("select count(*) from testtable"),1);
